add mod encode, decode
